+++ components/atomicComponents/modal/SnsIcon.tsx
