+++ components/moduleComponents/signIn/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SigninModule = () => {
+  return <div>signinModule</div>;
+};
+
+export default SigninModule;
