+++ components/moduleComponents/signUp/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SingUpModule = () => {
+  return (
+    <div>SingUpModule</div>
+  )
+}
+
+export default SingUpModule
