+++ pages/404.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const NotFound = () => {
+  return (
+    <div>잘못 들어옴</div>
+  )
+}
+
+export default NotFound
