+++ pages/folder.tsx
@@ -1,9 +1,9 @@
-import Profile from "../components/HeaderSection/Profile";
-import Footer from "../components/FooterSection/Footer";
-import LinkAdd from "../components/HeaderSection/LinkAdd";
-import FolderListBtn from "../components/folder/FolderListBtn";
-import Search from "../components/CardSection/Search";
-import styles from "../components/CardSection/CardSection.module.css";
+import Profile from "../components/moduleComponents/HeaderSection/Profile";
+import Footer from "../components/moduleComponents/FooterSection/Footer";
+import LinkAdd from "../components/moduleComponents/HeaderSection/LinkAdd";
+import FolderListBtn from "../components/moduleComponents/folder/FolderListBtn";
+import Search from "../components/moduleComponents/CardSection/Search";
+import styles from "../components/moduleComponents/CardSection/CardSection.module.css";
 import React from "react";
 
 export default function FolderPage() {
