+++ pages/shared.tsx
@@ -1,0 +1,13 @@
+import HeaderSection from "@/components/moduleComponents/HeaderSection/HeaderSection";
+import CardSection from "@/components/moduleComponents/CardSection/CardSection";
+import Footer from "@/components/moduleComponents/FooterSection/Footer";
+
+export default function Home() {
+  return (
+    <>
+      <HeaderSection />
+      <CardSection />
+      <Footer />
+    </>
+  );
+}
