+++ pages/signin.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import styles from "@/styles/signin.module.css";
+import SigninModule from "@/components/moduleComponents/signIn/index";
+
+const Singin = () => {
+  return (
+    <div className={styles.signin_back_color}>
+      <SigninModule />
+    </div>
+  );
+};
+
+export default Singin;
