+++ pages/signup.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import styles from "@/styles/signup.module.css";
+import SingUpModule from "@/components/moduleComponents/signUp/index";
+
+const Singup = () => {
+  return (
+    <div className={styles.signup_back_color}>
+      <SingUpModule />
+    </div>
+  )
+}
+
+export default Singup
